pyproject.toml
[tool.poetry]
name = "pd-dwi"
version = "1.1.3"
description = "Physiologically-Decomposed Diffusion-Weighted MRI machine-learning model for predicting response to neoadjuvant chemotherapy in invasive breast cancer"
authors = [
    "Maya Gilad <ms.maya.gilad@gmail.com>"
]
classifiers = [
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Development Status :: 3 - Alpha"
]
readme = "README.md"
homepage = "https://tcml-bme.github.io/"
repository = "https://github.com/TechnionComputationalMRILab/PD-DWI"
documentation = "https://pd-dwi.readthedocs.io/en/stable/"

[tool.poetry.scripts]
pd-dwi = "pd_dwi.scripts.cli:pd_dwi_cli"
pd-dwi-preprocessing = "pd_dwi.scripts.cli:preprocessing_cli"


[tool.poetry.dependencies]
python = ">=3.8.1,<3.9"
Cython = "^0.29.30"
jsonschema = "^4.6.0"
numpy = "^1.22.0"
pyradiomics = "^3.0.1"
PyYAML = "^6.0"
scikit-learn = "^1.0.2"
xgboost = "^1.6.1"
pandas = "^1.4.3"
pydantic = "^2.6.4"
pydantic-yaml = "^1.2.1"
StrEnum = {version = "^0.4.15", python = "<3.11"}
pydicom = "^2.4.4"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-subtests = "^0.12.1"
mypy = "^1.9.0"
types-PyYAML = "^6.0.12"
mkdocs = "^1.6.0"
mkdocs-plotly-plugin = "^0.1.3"
pymdown-extensions = "^10.8.1"
mkdocs-click = "^0.8.1"
mkdocs-material = "^9.5.28"
termynal = "^0.12.1"

[tool.poetry.extras]
preprocessing = ["pydicom"]

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "pd_dwi.scripts.*"
disallow_untyped_defs = false


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
